car images save2

+++ src/main/java/lk/acpt/riyapola/repo/CarRepo.java
@@ -3,5 +3,5 @@ package lk.acpt.riyapola.repo;
 import lk.acpt.riyapola.entity.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarRepo extends JpaRepository<Car,Long> {
+public interface CarRepo extends JpaRepository<Car,Integer> {
 }
